+++ frontend/app/components/atoms/Avatar.tsx
