--- frontend/app/components/molecules/ExperienceItem/index.ts
@@ -1,4 +1,0 @@
-import ExperienceItem, { ExperienceItemProps } from './ExperienceItem';
-
-export type { ExperienceItemProps };
-export default ExperienceItem;
