+++ frontend/app/components/molecules/UserInfo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Avatar from '../../atoms/Avatar/Avatar';
-import Text from '../../atoms/Typography/Text';
+import Avatar from '../atoms/Avatar';
+import Text from '../atoms/Typography/Text';
 import { Row, Col } from 'react-bootstrap';
 
 export interface UserInfoProps {
