+++ frontend/app/components/molecules/index.ts
@@ -1,14 +1,14 @@
 // Form molecule
-export { default as FormGroup } from './FormGroup/FormGroup';
+export { default as FormGroup } from './FormGroup';
 
 // IconWithText molecule
-export { default as IconWithText } from './IconWithText/IconWithText';
+export { default as IconWithText } from './IconWithText';
 
 // UserInfo molecule
-export { default as UserInfo } from './UserInfo/UserInfo';
+export { default as UserInfo } from './UserInfo';
 
 // AnimatedCard molecule
-export { default as AnimatedCard } from './AnimatedCard/AnimatedCard';
+export { default as AnimatedCard } from './AnimatedCard';
 
 // ExperienceItem molecule
 // export { default as ExperienceItem } from './ExperienceItem/ExperienceItem';
@@ -18,3 +18,8 @@ export { default as AnimatedCard } from './AnimatedCard/AnimatedCard';
 
 // // types
 // export type { ExperienceItemProps } from './ExperienceItem/ExperienceItem';
+
+import ExperienceItem, { ExperienceItemProps} from "@/app/components/molecules/ExperienceItem";
+
+export type { ExperienceItemProps };
+export default ExperienceItem;
