--- frontend/app/components/organisms/Experience/index.ts
@@ -1,3 +1,0 @@
-import Experience from './Experience';
-
-export default Experience;
