+++ frontend/app/components/organisms/index.ts
@@ -1,7 +1,12 @@
 // Layout organisms
-export { default as Header } from './Header/Header';
-export { default as Footer } from './Footer/Footer';
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
 
 // Section organisms
-export { default as Banner } from './Banner/Banner';
-export { default as ContactForm } from './ContactForm/ContactForm';
+export { default as Banner } from './Banner';
+export { default as ContactForm } from './ContactForm';
+import Experience from "@/app/components/organisms/Experience";
+export { default as ExperienceItem } from './ExperienceItem';
+
+export default Experience;
+
