+++ frontend/app/components/templates/MainLayout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from '../../organisms/Header/Header';
-import Footer from '../../organisms/Footer/Footer';
+import Header from '../organisms/Header';
+import Footer from '../organisms/Footer';
 
 interface MainLayoutProps {
   children: React.ReactNode;
